More intelligent algorithm for JS-DOM getAttrs.
The JS-DOM getAttrs logic was ignoring data-type in the return values, and
always returning a string.  The new logic allows a user to specify whether
the output value should be coerced to a specific type (string, boolean,
number) when read from a DOM node based on the type of the parameter
passed to the object.

In order to retrieve a string, the attribute name should be a string.  In
order to retrieve a boolean, then a boolean should be passed in.  In order
to retrieve a number, then a number should be passed in.  All other values
will be interpreted as a string before being returned.