Integration from v0.1.0 to v0.2.0
This merge completes the staging from Pull Request #63.  It includes the
introduction of the JS-DType and JS-Storage modules into the API; basic
user and group support for the JS-SP (SharePoint) module; improves the
efficiency of the JS-SP Query's Sum algorithm; improved the handling of
non-integer rational numbers in JS-SP Date Arithmatic; improves the logic
underlying the getAttributes and exportAttributes functions in the JS-DOM
module; cleans up the get and export interface throughout JSpyder;
corrects several documentation deficiencies; and includes a few bug fixes.

In addition to these changes, this version includes an explicit version
integration, a re-compiled documentation folder, some basic cleanup prior
to the "real" version integration, and has been re-packaged for deployment
into the "jspyder" folder for easier distribution.